scripts/requirements.txt
firebase-admin>=6.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
feedparser>=6.0.10
python-dotenv>=1.0.0

